+++ src/gui/GuiWindow.cpp
@@ -1,11 +1,11 @@
 #include "GuiWindow.h"
 
 GuiWindow::GuiWindow(std::string guiName, Gui &parent, sf::Vector2f position, sf::Vector2f size, sf::Vector2f origin,
-	sf::Color backgroundColour, sf::Color hoverBackgroundColour, sf::Color borderColour, int borderThickness) : ChildGui(guiName, parent, position, size, origin) {
+	sf::Color backgroundColour, sf::Color hoverBackgroundColour, sf::Color borderColour, float borderThickness) : ChildGui(guiName, parent, position, size, origin) {
 	setupBackground(backgroundColour, hoverBackgroundColour, borderColour, borderThickness);
 }
 
-void GuiWindow::setupBackground(sf::Color colourBackground, sf::Color colourBackgroundHover, sf::Color borderColour, int borderThickness) {
+void GuiWindow::setupBackground(sf::Color colourBackground, sf::Color colourBackgroundHover, sf::Color borderColour, float borderThickness) {
 	backgroundColour = colourBackground;
 	hoverBackgroundColour = colourBackgroundHover;
